services/vanna/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
psycopg[binary]==3.1.18
asyncpg==0.29.0
python-dotenv==1.0.0
groq==0.4.1
httpx==0.25.2
sqlparse==0.4.4


